Map Provider
TODO: Set consumer provider structure to map screen by using UserModel as provider.
